pyproject.toml
[project]
name = "weather-epaper"
version = "0.1.0"
description = "Weather display for Waveshare 2.7\" e-paper on Raspberry Pi"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
pi = [
    "spidev>=3.5",
    "gpiozero>=2.0",
]

[project.scripts]
weather-epaper = "weather_epaper.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src/weather_epaper", "third_party/waveshare"]

[tool.hatch.build.targets.wheel.sources]
"src/weather_epaper" = "weather_epaper"
"third_party/waveshare" = "waveshare"

[tool.hatch.build.targets.sdist]
include = ["src/weather_epaper", "third_party/waveshare", "pyproject.toml"]
